Added instance methods to numerics and filesizes.
...such as .to?B and .?B.
