+++ lib/ext/numeric.rb
@@ -1,0 +1,18 @@
+class Numeric
+  ( [ '' ] + 'kMGTPEZY'.chars ).map do |magnitude|
+    [ "#{magnitude}B", "#{magnitude}iB" ]
+  end.flatten.map(&:to_sym).each do |prefix|
+    # Multiply
+    define_method prefix do
+      Filesize.from("#{to_i} #{prefix}")
+    end
+  end.map { |prefix| "to_#{prefix}" }.map do |prefix|
+    prefix.gsub(/([a-z])_([A-Z])/, '\1\2')
+  end.map(&:to_sym).each do |method|
+    prefix = method.to_s.gsub(/^to_?/, '')
+    # Divide
+    define_method method do
+      Filesize.from("#{to_i} B").to(prefix)
+    end
+  end
+end
